confirmação funcionando

+++ src/app/residentes/residentes.component.ts
@@ -1,3 +1,4 @@
+import { DialogConfirmService } from './dialog-confirm.service';
 import { ActivatedRoute } from '@angular/router';
 import { Component, OnInit } from '@angular/core';
 import { ResidentesService } from './residentes.service';
@@ -12,7 +13,7 @@ export class ResidentesComponent implements OnInit {
 
   residentes: Residente[]
 
-  constructor(private residentesService: ResidentesService) { }
+  constructor(private residentesService: ResidentesService, private dialogConfirmService: DialogConfirmService) { }
 
   ngOnInit() {
     this.residentesService.residentes()
@@ -20,8 +21,14 @@ export class ResidentesComponent implements OnInit {
   }
 
 
-  deleteResidente(id: string){
-    this.residentesService.deleteResidente(id)
-    .subscribe(result => this.residentesService.residentes().subscribe(residentes => this.residentes = residentes))
+  deleteResidente(id: string): void {
+    this.dialogConfirmService.confirm(`Deseja excluir o residente?`)
+      .then((isTrue) => {
+        if (isTrue) {
+          this.residentesService.deleteResidente(id)
+            .subscribe(result => this.residentesService.residentes()
+              .subscribe(residentes => this.residentes = residentes))
+        }
+      })
   }
 }
